MOS-59 change timing info test to use uint32_t for indices

+++ test/TimingInfoTest.cpp
@@ -17,7 +17,7 @@ TEST_CASE("TimingInfoTest", "[core]")
     {
         isx::Time start;
         isx::Ratio step(50, 1000);
-        uint64_t numTimes = 20;
+        uint32_t numTimes = 20;
         isx::TimingInfo timingInfo(start, step, numTimes);
         REQUIRE(timingInfo.getStart() == start);
         REQUIRE(timingInfo.getStep() == step);
@@ -28,7 +28,7 @@ TEST_CASE("TimingInfoTest", "[core]")
     {
         isx::Time start;
         isx::Ratio step(50, 1000);
-        uint64_t numTimes = 20;
+        uint32_t numTimes = 20;
         isx::TimingInfo timingInfo(start, step, numTimes);
         isx::Time expected = start.addSecs(1);
         REQUIRE(timingInfo.getEnd() == expected);
@@ -38,7 +38,7 @@ TEST_CASE("TimingInfoTest", "[core]")
     {
         isx::Time start;
         isx::Ratio step(50, 1000);
-        uint64_t numTimes = 20;
+        uint32_t numTimes = 20;
         isx::TimingInfo timingInfo(start, step, numTimes);
         REQUIRE(timingInfo.getDuration() == 1);
     }
@@ -48,7 +48,7 @@ TEST_CASE("TimingInfoTest", "[core]")
         isx::Time start;
         isx::Ratio step(50, 1000);
         isx::TimingInfo timingInfo(start, step, 100);
-        uint64_t index = timingInfo.convertTimeToIndex(start);
+        uint32_t index = timingInfo.convertTimeToIndex(start);
         REQUIRE(index == 0);
     }
 
@@ -58,7 +58,7 @@ TEST_CASE("TimingInfoTest", "[core]")
         isx::Ratio step(50, 1000);
         isx::TimingInfo timingInfo(start, step, 100);
         isx::Time time = start.addSecs(-2);
-        uint64_t index = timingInfo.convertTimeToIndex(time);
+        uint32_t index = timingInfo.convertTimeToIndex(time);
         REQUIRE(index == 0);
     }
 
@@ -68,7 +68,7 @@ TEST_CASE("TimingInfoTest", "[core]")
         isx::Ratio step(50, 1000);
         isx::TimingInfo timingInfo(start, step, 100);
         isx::Time time = start.addSecs(isx::Ratio(25, 1000));
-        uint64_t index = timingInfo.convertTimeToIndex(time);
+        uint32_t index = timingInfo.convertTimeToIndex(time);
         REQUIRE(index == 0);
     }
 
@@ -78,7 +78,7 @@ TEST_CASE("TimingInfoTest", "[core]")
         isx::Ratio step(50, 1000);
         isx::TimingInfo timingInfo(start, step, 100);
         isx::Time time = start.addSecs(isx::Ratio(50, 1000));
-        uint64_t index = timingInfo.convertTimeToIndex(time);
+        uint32_t index = timingInfo.convertTimeToIndex(time);
         REQUIRE(index == 1);
     }
 
@@ -88,7 +88,7 @@ TEST_CASE("TimingInfoTest", "[core]")
         isx::Ratio step(50, 1000);
         isx::TimingInfo timingInfo(start, step, 100);
         isx::Time time = start.addSecs(isx::Ratio(74, 1000));
-        uint64_t index = timingInfo.convertTimeToIndex(time);
+        uint32_t index = timingInfo.convertTimeToIndex(time);
         REQUIRE(index == 1);
     }
 
@@ -98,7 +98,7 @@ TEST_CASE("TimingInfoTest", "[core]")
         isx::Ratio step(50, 1000);
         isx::TimingInfo timingInfo(start, step, 100);
         isx::Time time = timingInfo.getEnd();
-        uint64_t index = timingInfo.convertTimeToIndex(time);
+        uint32_t index = timingInfo.convertTimeToIndex(time);
         REQUIRE(index == 99);
     }
 
@@ -106,7 +106,7 @@ TEST_CASE("TimingInfoTest", "[core]")
     {
         isx::TimingInfo timingInfo;
         isx::Time time = timingInfo.getEnd().addSecs(2);
-        uint64_t index = timingInfo.convertTimeToIndex(time);
+        uint32_t index = timingInfo.convertTimeToIndex(time);
         REQUIRE(index == 99);
     }
